tests/requirements.txt
pytest==2.7.3
pytest-selenium
marionette_driver==1.3.0
